Change data and code #156

+++ descriptive/events_count/code/build.R
@@ -8,21 +8,23 @@ out_file <- '../output/event_counts.txt'
 
 main <- function() {
   files_dir <-
-    list.files(in_samples, pattern = ".csv", full.names = T)[-2] # Ignore all_zipcode_months (too big)
-  file.remove(out_file)
+    list.files(
+      '../../../drive/derived_large/min_wage',
+      pattern = "statutory_mw.csv",
+      full.names = T
+    )
+  if (file.exists(out_file)) file.remove(out_file)
   
-  for (dd in files_dir) {
-    file_name <- stringr::str_remove(dd, '.*/')
+  states <- c("zipcode", "county", "state") # Add local later
+  
+  for (ss in states) {
+    if (ss != "zipcode") data <- fread(files_dir[1])
+    else data <- fread(files_dir[2])
     
-    data <- fread(dd)
+    events <- get(paste0('count_', ss, '_events'))(data)
     
-    data[, mw_event := fifelse(actual_mw != shift(actual_mw), 1, 0), by =
-           county_num]
-    
-    n <- sum(data$mw_event == 1, na.rm = T)
-    
-    text <- paste0('Number of MW changes in ', file_name, ": ",
-                   n)
+    text <- paste('Number of', ss, 'MW events:',
+                  events)
     
     write.table(
       text,
@@ -35,6 +37,50 @@ main <- function() {
     
     remove(data)
   }
+}
+
+count_state_events <- function(data) {
+  data[,
+       countyfips := fifelse(nchar(countyfips) == 4,
+                             paste0(0, countyfips),
+                             as.character(countyfips))][,
+                                                        state := substr(countyfips, 1, 2)]
+  
+  data_state <- data[,
+                     .(state_mw = unique(state_mw)),
+                     by = list(month, year, state)]
+  
+  data_state[, state_event := fifelse(state_mw > shift(state_mw), 1, 0), by =
+               state]
+  
+  n <- sum(data_state$state_event, na.rm = T)
+  
+  return(n)
+  remove(data_state)
+}
+
+count_zipcode_events <- function(data) {
+  data[, zipcode_event := fifelse(actual_mw > shift(actual_mw), 1, 0), by =
+         zipcode]
+  
+  n <- sum(data$zipcode_event, na.rm = T)
+  
+  return(n)
+}
+
+count_county_events <- function(data) {
+  # Change actual_mw to county_mw when it becomes available
+  
+  data[, county_event := fifelse(actual_mw > shift(actual_mw), 1, 0), by =
+         countyfips]
+  
+  n <- sum(data$county_event, na.rm = T)
+  
+  return(n)
+}
+
+count_local_events <- function (data) {
+  
 }
 
 # Execute
